fix: resolve Flutter analyzer findings

+++ app/lib/features/system/views/backup_page.dart
@@ -23,7 +23,7 @@ bool isSupportedBackupFilename(String filename) {
 }
 
 Duration restoreProgressRetryDelay(int failures) {
-  final exponent = failures.clamp(0, 4) as int;
+  final exponent = failures.clamp(0, 4);
   return Duration(seconds: 1 << exponent);
 }
